scripts_python/pyproject.toml
[project]
name = "scripts-python"
version = "0.1.0"
description = "Python scripts for all-the-benchmarks"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.3.1",
    "pytest>=8.4.1",
    "pyyaml>=6.0",
    "scipy>=1.14.0",
]
